JOTM -113/09/2018 Coloca lectura p5000 y porcentajes de pipas lecturas
Agrega en lectura código para que al inicializar si es lectura coloque por defecto un valor extraído del servicio en la lectura del P5000 y de los porcentajes de medidor.

+++ 1.Layers/1.1Presentation/1.1.2Mobile/app/src/main/java/com/example/neotecknewts/sagasapp/Model/AlmacenDTO.java
@@ -20,7 +20,7 @@ public class AlmacenDTO {
     private double PorcentajeMedidor;
 
     @SerializedName("CantidadP5000")
-    private double CantidadP5000;
+    private int CantidadP5000;
 
     @SerializedName("IdTipoMedidor")
     private int IdTipoMedidor;
@@ -56,11 +56,11 @@ public class AlmacenDTO {
         PorcentajeMedidor = porcentajeMedidor;
     }
 
-    public double getCantidadP5000() {
+    public int getCantidadP5000() {
         return CantidadP5000;
     }
 
-    public void setCantidadP5000(double cantidadP5000) {
+    public void setCantidadP5000(int cantidadP5000) {
         CantidadP5000 = cantidadP5000;
     }
